support for setting value to undefined
I can not find the way, how to set value to undefined using freezer. (I dont want to remove the key in hash, but set the key to undefined.)

'''javascript
parent.set("myKey",undefined);
'''
I would be able to do it with this fix. Can you please check it, that this change will not eventually break something else.

Thanks.